+++ src/python/src/openmeta_module.cc
@@ -6527,6 +6527,53 @@ translate_capture_rational_metadata_document(
 }
 
 static std::shared_ptr<PyDocument>
+translate_flash_metadata_document(
+    std::shared_ptr<PyDocument> source,
+    MetadataCaptureTranslationSourceMode source_mode,
+    MetadataCaptureTranslationConflictPolicy conflict_policy,
+    uint32_t max_added_entries, uint32_t max_source_properties,
+    uint32_t max_operations, uint32_t max_text_bytes_per_property,
+    uint64_t max_total_text_bytes)
+{
+    MetadataFlashTranslationOptions options;
+    options.source_mode                 = source_mode;
+    options.conflict_policy             = conflict_policy;
+    options.max_source_properties       = max_source_properties;
+    options.max_added_entries           = max_added_entries;
+    options.max_operations              = max_operations;
+    options.max_text_bytes_per_property = max_text_bytes_per_property;
+    options.max_total_text_bytes        = max_total_text_bytes;
+
+    MetaStore translated;
+    MetadataCaptureTranslationResult result;
+    {
+        nb::gil_scoped_release gil_release;
+        result = translate_xmp_flash_metadata(source->store, options,
+                                              &translated);
+    }
+    if (result.status != MetadataCaptureTranslationStatus::Ok) {
+        std::string message = "metadata flash translation failed: ";
+        message += metadata_capture_translation_status_name(result.status);
+        if (result.failed_mapping != MetadataCaptureTranslationMapping::None) {
+            message += " for ";
+            message += metadata_capture_translation_mapping_name(
+                result.failed_mapping);
+        }
+        if (result.failed_source_entry != kInvalidEntryId) {
+            message += " at source entry ";
+            message += std::to_string(result.failed_source_entry);
+        }
+        throw std::invalid_argument(message);
+    }
+
+    auto document                        = std::make_shared<PyDocument>();
+    document->store                      = std::move(translated);
+    document->result.xmp.entries_decoded = active_xmp_entry_count(
+        document->store);
+    return document;
+}
+
+static std::shared_ptr<PyDocument>
 translate_image_geometry_document(
     std::shared_ptr<PyDocument> source,
     const TransferTargetImageSpec& target_image_spec,
@@ -8980,8 +9027,18 @@ NB_MODULE(_openmeta, m)
         .value("XmpExposureIndex",
                MetadataCaptureTranslationMapping::XmpExposureIndex)
         .value("XmpFlashEnergy",
-               MetadataCaptureTranslationMapping::XmpFlashEnergy);
+               MetadataCaptureTranslationMapping::XmpFlashEnergy)
+        .value("XmpFlash", MetadataCaptureTranslationMapping::XmpFlash);
 
+
+    m.attr("METADATA_FLASH_TRANSLATION_CONTRACT_VERSION") = nb::int_(
+        kMetadataFlashTranslationContractVersion);
+    m.attr("METADATA_FLASH_TRANSLATION_MAX_ADDED_ENTRIES") = nb::int_(
+        kMetadataFlashTranslationMaxAddedEntries);
+    m.attr("METADATA_FLASH_TRANSLATION_MAX_SOURCE_PROPERTIES") = nb::int_(
+        kMetadataFlashTranslationMaxSourceProperties);
+    m.attr("METADATA_FLASH_TRANSLATION_MAX_TOTAL_TEXT_BYTES") = nb::int_(
+        kMetadataFlashTranslationMaxTotalTextBytes);
 
     m.attr("METADATA_CAPTURE_RATIONAL_TRANSLATION_CONTRACT_VERSION") = nb::int_(
         kMetadataCaptureRationalTranslationContractVersion);
@@ -9022,8 +9079,11 @@ NB_MODULE(_openmeta, m)
                MetadataCaptureTranslationStatus::EntryLimitExceeded)
         .value("OperationLimitExceeded",
                MetadataCaptureTranslationStatus::OperationLimitExceeded)
-        .value("InternalError",
-               MetadataCaptureTranslationStatus::InternalError);
+        .value("InternalError", MetadataCaptureTranslationStatus::InternalError)
+        .value("IncompleteSource",
+               MetadataCaptureTranslationStatus::IncompleteSource)
+        .value("UnsupportedSourceShape",
+               MetadataCaptureTranslationStatus::UnsupportedSourceShape);
 
     m.attr("METADATA_CAPTURE_TRANSLATION_MAX_ADDED_ENTRIES") = nb::int_(
         kMetadataCaptureTranslationMaxAddedEntries);
@@ -10030,6 +10090,18 @@ NB_MODULE(_openmeta, m)
              = kMetadataCaptureTranslationMaxTextBytesPerProperty,
              "max_total_text_bytes"_a
              = kMetadataCaptureRationalTranslationMaxTotalTextBytes)
+        .def("translate_flash_metadata", &translate_flash_metadata_document,
+             "source_mode"_a = MetadataCaptureTranslationSourceMode::DirtyOnly,
+             "conflict_policy"_a
+             = MetadataCaptureTranslationConflictPolicy::FailOnConflict,
+             "max_added_entries"_a = kMetadataFlashTranslationMaxAddedEntries,
+             "max_source_properties"_a
+             = kMetadataFlashTranslationMaxSourceProperties,
+             "max_operations"_a = kMetadataCaptureTranslationMaxOperations,
+             "max_text_bytes_per_property"_a
+             = kMetadataCaptureTranslationMaxTextBytesPerProperty,
+             "max_total_text_bytes"_a
+             = kMetadataFlashTranslationMaxTotalTextBytes)
         .def("translate_image_geometry", &translate_image_geometry_document,
              "target_image_spec"_a,
              "source_mode"_a = MetadataGeometryTranslationSourceMode::DirtyOnly,